requirements.txt
boto3==1.4.7
botocore==1.7.1
certifi==2017.7.27.1
chardet==3.0.4
docutils==0.14
idna==2.6
jmespath==0.9.3
PyJWT==1.5.2
PySocks==1.6.7
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.10
six==1.10.0
twilio==6.5.2
urllib3==1.22
